refactor(test) : 로또 객체  수정에 따른 test 코드 수정
