+++ src/test/java/lotto/LottosTest.java
@@ -1,10 +1,14 @@
-package lottogame;
+package lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
 import java.util.Map;
+import lotto.domain.game.Lotto;
+import lotto.domain.statistics.LottoMatch;
+import lotto.domain.game.Lottos;
+import lotto.domain.game.WinningNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 public class LottosTest {
@@ -16,9 +20,9 @@ public class LottosTest {
         List<Integer> list2 = List.of(3, 5, 11, 16, 32, 38);
         List<Integer> list3 = List.of(7, 11, 16, 35, 36, 44);
 
-        Lotto lotto1 = new Lotto(list1);
-        Lotto lotto2 = new Lotto(list2);
-        Lotto lotto3 = new Lotto(list3);
+        Lotto lotto1 = new Lotto(list1, false);
+        Lotto lotto2 = new Lotto(list2, false);
+        Lotto lotto3 = new Lotto(list3, false);
         List<Lotto> lottos = List.of(lotto1, lotto2, lotto3);
 
         assertDoesNotThrow(() -> new Lottos(lottos));
@@ -31,9 +35,9 @@ public class LottosTest {
         List<Integer> list2 = List.of(3, 5, 11, 16, 32, 38);
         List<Integer> list3 = List.of(7, 11, 16, 35, 36, 44);
 
-        Lotto lotto1 = new Lotto(list1);
-        Lotto lotto2 = new Lotto(list2);
-        Lotto lotto3 = new Lotto(list3);
+        Lotto lotto1 = new Lotto(list1, false);
+        Lotto lotto2 = new Lotto(list2, false);
+        Lotto lotto3 = new Lotto(list3, false);
 
         Lottos lottos = new Lottos(List.of(lotto1, lotto2, lotto3));
